Make dbcp optional
Allow user to enable/disable DBCP. By default it is not required as it
is only useful when there is huge number of database connections
involved. Also in some scenarios like when pgpool is running in front of
PostgreSQL DBCP is creating issues. There is no need for client side
pool in such cases.