Eliminate POST and DEL /job race conditions
For both these routes, there was a race condition when you immediately
try and check the status of a job after starting or stopping it.

For example, if you POST'ed an async job and then immediately try and
do a GET on that jobId, the jobInfoActor may not have persisted the job
yet, so you would get a 404. Similiar could happen for stopping a job,
if you immediately tried to GET after cancelling, it may still show as
  RUNNING