test(jobserver): Add testcases for AkkaClusterSupervisor (#1029)
Steps:
- Create one akka cluster for master at port 2552
- Create slave akka cluster at random port
- Create a fake jobmanager actor
- Override launchDriver method and instead of
creating a new JVM, create a cluster and JobManagerActor
in the same process. This newly created actor is
used for communication with master

Since, creating a cluster is expensive, we create the
cluter only once and reuse it for each testcase.
All the contexts that we create during a testcase, we
remove them in after{} block afterwards. This makes
sure that each testcase has a clean start.