feat(jobserver): Remove extra states from contextInitInfos (#1055)
Since, we have a DB table for context now, we don't
need to put the context config and is-adhoc flag in
the hashmap anymore. Now, we pull this information from
DB.

This hashmap also contains the success and failure
callbacks. It doesn't make sense to put them in DB
because they are just temporary. If somehow the master
crashes, the user who is waiting for the response
won't get the response anyway.

Change-Id: I7290427d06a8ff52d5e4655ba0bf054ad8306b47