feat(jobserver): Add watch on existing context upon SJS restart
AkkaClusterSupervisorActor adds a watch when a new context
is launched. This watch is used to catch the Terminated
event whenever the context JVM is disconected.

Due to SJS Master restart, AkkaClusterSupervisorActor can
lose all the watches. This patch adds a check on SJS startup
to figure out which contexts JVMs are reconnectable and send
then a message to AkkaClusterSupervisorActor which contains
the related ActorRef. AkkaClusterSupervisorActor then adds a
watch to old JVMs and is able to receive the Terminated
event again.

Having the watch is important for proper cleanup. Without
having the watch it could be that:
1- A user stops a context, the DB will have STOPPING state
but the final state FINISHED will not be written.
2- The context gets killed unexpectedly, then
* For supervise mode, RESTARTING won't be written in DB.
* Without supervise mode, KILLED won't be written.