fix(jobserver): At SJS startup, handle timeout exception
At startup, jobserver looks at all the contexts which
were in RUNNING/STOPPING state and tries to reconnect
with them. If reconnection succeeds then jobserver
tries to join the existing Akka cluster otherwise it
writes error state for that context.

But there can be a case where a context/driver exists
but doesn't respond within the timeout (3 seconds in
this case) and a Timeout exception is thrown from
system.actorSelection(address).resolveOne(timeout).
This case was not handled before which leads to
jobserver exiting the JVM.

This change now catches all the expected exceptions and
handles them gracefully.

Change-Id: Ic90d5bdc8efe893e9238225f6d6017d551ffc815