Multiple fixes for jobserver (#1126)
* fix(jobserver): At startup, cleanup entities in Stopping state

Logic:
Try to resolve the reference to driver, if connection is
OK, driver is up and don't do anything. Otherwise, update DAO
with error state for context and job.

Change-Id: I89c48bdd307e446b5dd1ae736314ef1e2e8c9347

* fix(jobserver): Fix NPE on JobServer startup

Change-Id: I694595aa0c887ef1227945d840d21d9f6636e19a

* fix(jobserver): Set ERROR for context if restart fails

This is a fix for the problem mentioned in the commit
message of "Handle delete properly....." change.

Within the restart logic, if something goes wrong with
DAO, we kill ourself. This raises a Terminated event on
master which saves RESTARTING in DAO whereas the slave
is never going to come back because we noramlly killed
it.

This fix tries to update the state in DAO before doing
a PoisonPill. This keeps the states clean and also
takes the burden from Terminated event because now less
scenario can lead to RESTARTING state.

Change-Id: I563d278393925a99704ae9f92b6b28528b118ad0