fix(jobserver): Do not convert stream to character stream
Previously we only had 1 usecase for HdfsFSFacade.get()
method, to read the config file. So, the function was
converting the stream returned from Hadoop to a
character stream.

The new usecase is to read jar files also with the same
function but the jars files are binaries and converting
binary to a character stream and then storing it locally
leads to jar being corrupt and unusable.

The fix is to let the client decide what to do with
the stream.

Invalid inputs found at:
https://www.cl.cam.ac.uk/~mgk25/ucs/examples/UTF-8-test.txt
https://stackoverflow.com/a/41167392/1470127

Change-Id: I47a524a634e29af678e59c78a58c7ff458314d0d