fix(tests): Fix curator framework import and namespace
Apache-curator import seem not to work correctly and not to resolve
all dependencies. The project is compiling and running because
curator-framework is also required by spark-core. As spark-core
uses an older version of curator, import framework explicitly once
again (curator-framework has curator-clients and ZK as dependency).

As Curator 4.0.1 is not a dependency for all projects, we still see
curator 2.6.0 in subprojects like akka-app and jobserver-tests.
Eclipse seems to pick up this dependency and use curator 2.6.0
testing even other projects. After checking ivy reports, testing
assembly and with InteliJ it seems to be only a problem for Eclipse.

Namespace should not start with / starting from curator version 2.7.

TODO: The issue should be fixed and tested more properly later.
Ideal solution will be to use curator 4.0.1 with ZK 3.5.
Test library can be also updated then (TestServer function
in 4.0.1 is incompatible with ZK 3.4)

Change-Id: I87f2220a69f9f0821252f0f8aac4efe480e0bee7