feat(jobserver): Make AkkaClusterSupervisorActor a singleton
* Create AkkaClusterSupervisorActor as an akka cluster singleton
  on jobserver startup
* Adapt application.conf for a cluster singleton
* Adapt JobserverSpec to also check for singleton infrastructure
  in case of context-per-jvm=true
* Temporarily adapt changes to jobManager zombie kill features as
  long as we have no split brain resolver.

In jobserver HA mode the logic following a MemberUp event
(initializing a context) or Terminated event (updating context
and associated job states in the database) can conflict when
handled by two or more active jobservers. Moving this logic to a
cluster singleton ensures that this logic is executed exclusively
by one actor.
See: https://doc.akka.io/docs/akka/current/cluster-singleton.html

Making the AkkaClusterSupervisorActor a singleton is a relatively
simple, straight-forward solution, which requires few changes and
is therefore a good initial solution for HA. The downside is
that a singleton now handles more than just the necessary
AkkaClusterEvents (i.e. context management parts, which could be
handled locally by each jobserver), which represents a potential
bottleneck in the future.

A recommended optimization for the future would be to refactor
the AkkaClusterSupervisorActor into two actors (cluster and
context managing parts), where only the cluster parts will be
handled by a singleton. To not overburden the initial HA of
jobserver, this refactoring is left out for now.

Change-Id: I81fc56a76f49ed238bbc7781f1ebb0ef85fb7d7e