fix(jobserver): Decrease down-removal-margin
Set custom-downing.down-removal-margin to 1s instead of 10s.
The margin effectively determines how much time elapses between
SBR downing unreachable nodes and instantiating a new singleton
actor on the second jobserver if the first one was downed.

Messages sent to an unreachable singleton will be lost,
ultimately leading to a timeout of the initial API operation.
But Singleton messages being sent during the down-removal-margin
will be buffered by the singleton proxy and eventually processed
by the singleton as soon as it is created.
This corner case might lead to timeouts and seemingly failed
operations that actually succeeded in the backend (e.g context
creation/deletion).

It seems like a very low down-removal-margin in our case is
acceptable, as a) the risk of having two singletons is very
low at all,  b) there are currently only very few messages
that the singleton has to handle exclusively and c) it is
practically not possible that these messages (e.g. MemberUp
and Terminated) are handled by both singletons (as they don't
share the same cluster anymore).
The therefore avoided effect of having seemingly failed
operations would probably be worse.

Change-Id: Id5d8b7cd90b09e1cd7372b184b4c6e1766c3d106