fix(test): Fix flaky test in JobManagerActorSpec
Very often the addressed testcase in JobManagerActorSpec
would fail with error

java.lang.AssertionError: assertion failed:
expected StartJob Received,
found JobRestartFailed(
  jobId1,spark.jobserver.util.NoJobConfigFoundException:
  Failed to load config for job with id: jobId1)

This patch fixes that and the reason was that
the underlying dao query which fetches the jobs for
a context has no defined order. The testcase has
2 jobs namely jobId and jobId1. If the order was
[jobId, jobId1], the test case worked fine because
"StartJob Received" would be received first followed
by StartJob and JobRestartFailed. If the order changes
then the messages received are JobRestartFailed,
"StartJob Recieved" and StartJob.

This line in the code
spyProbe.expectMsg("StartJob Received") made the
order important.

Change-Id: I30fd4fad35ca8399d3761093f4190afc63d63e4a