feat(jobserver): Add URI BinaryType
This change allows binaries specified through URIs
in the cp parameters (or in dependent-jar-uris) to have
a proper BinaryType counterpart in the future.

URI BinaryType is not reflected in fromMediaType
function, because this type is created for an internal
(not visible to the end-user) use and should never be
uploaded or read by the end user.

Change-Id: I01c76dec533305c81408562935b47870750a3e19