feat(jobserver): Extend JobInfo to maintain cp BinaryInfo list
Move cp (short for "class path", which is a list of binaries
to start application from) from JobConfig (put there for
convenience reasons in previous changes) to JobInfo object.
This is more correct from design perspective as information
belongs to JobInfo and no additional job config retrievals
are necessary.

In this change:
* Add Seq[BinInfo] parameter to JobInfo object.
* Add new helper for BinaryManager to parse user input
already in WebApi and be able to access parsed BinaryInfo
in every scenario (start/restart of a job).
* Fix JsonProtocols for Zookeeper. NOTE: SQL, Cassandra
DAOs compile, but have not yet been adapted to save
and get extended JobInfo object.
* Add various tests for the new feature.
* Many old tests are touched in this change, but mostly
it is due to change in JobStart message format.

Change-Id: Ibe0c18042af90c155d44141f8d44b0fb6d18b400